feat: added provideToastr function and used it in ToastrModule for backwards compatibility
Made it return EnvironmentProviders in order to not be misused in component providers but only on bootstrapApplication providers or route providers
