+++ src/lib/toastr/toastr.module.ts
@@ -2,11 +2,7 @@ import { ModuleWithProviders, NgModule } from '@angular/core';
 
 import { Toast } from './toast.component';
 import { DefaultNoComponentGlobalConfig, GlobalConfig, TOAST_CONFIG } from './toastr-config';
-
-export const DefaultGlobalConfig: GlobalConfig = {
-  ...DefaultNoComponentGlobalConfig,
-  toastComponent: Toast,
-};
+import { provideToastr } from './toast.provider';
 
 @NgModule({
   imports: [Toast],
@@ -16,15 +12,7 @@ export class ToastrModule {
   static forRoot(config: Partial<GlobalConfig> = {}): ModuleWithProviders<ToastrModule> {
     return {
       ngModule: ToastrModule,
-      providers: [
-        {
-          provide: TOAST_CONFIG,
-          useValue: {
-            default: DefaultGlobalConfig,
-            config,
-          },
-        },
-      ],
+      providers: [provideToastr(config)],
     };
   }
 }
